Feat: remove the duplicate command
